+++ src/conditional-visibility.ts
@@ -28,7 +28,7 @@ import API from './module/api';
 /* ------------------------------------ */
 /* Initialize module					*/
 /* ------------------------------------ */
-Hooks.once('init', async function () {
+Hooks.once('init', function () {
   log(' init ' + CONSTANTS.MODULE_NAME);
   // Assign custom classes and constants here
 
@@ -36,7 +36,7 @@ Hooks.once('init', async function () {
   registerSettings();
 
   // Preload Handlebars templates
-  await preloadTemplates();
+  preloadTemplates();
 
   // Register custom sheets (if any)
   initHooks();
@@ -52,7 +52,7 @@ Hooks.once('setup', function () {
 /* ------------------------------------ */
 /* When ready							*/
 /* ------------------------------------ */
-Hooks.once('ready', async function () {
+Hooks.once('ready', function () {
   if (!game.modules.get('lib-wrapper')?.active && game.user?.isGM) {
     let word = 'install and activate';
     if (game.modules.get('lib-wrapper')) word = 'activate';
@@ -123,6 +123,14 @@ Hooks.once('libChangelogsReady', function () {
     CONSTANTS.MODULE_NAME,
     'levels',
     `With levels module enabled and active, **if the scene is with "Token vision" set to false (unchecked box)**, after selected a token and click on the canvas with the option "Release on left click" enable the hidden token are visible for a small instant this is a incompatibility with the [Levels](https://github.com/theripper93/Levels) module i cannot solve, the simple solution is just enable the token vision on the current scene.`,
+    'minor',
+  );
+
+  //@ts-ignore
+  libChangelogs.registerConflict(
+    CONSTANTS.MODULE_NAME,
+    'less-fog',
+    `With less-fog module enabled and active, The module "less fog" breaks the dm view of tokens. The gm still see an invisible token as other tokens, but the players don't so is a minor issue. The solution is just make sure the module 'Less Fog' settings panel 'Reveal Tokens' is false (uncheked box), is itoptional but advisable to set 'Reveal to All Players' is false (uncheked box) either.`,
     'minor',
   );